+++ src/main/java/com/ecommerce/fruitstore/domain/OrderSummary.java
@@ -1,5 +1,6 @@
 package com.ecommerce.fruitstore.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -10,7 +11,7 @@ public class OrderSummary {
 
     private int oranges;
 
-    private double totalPrice;
+    private BigDecimal totalPrice;
 
     private LocalDateTime orderDateTime;
 
@@ -26,7 +27,7 @@ public class OrderSummary {
         return oranges;
     }
 
-    public double getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
@@ -34,7 +35,7 @@ public class OrderSummary {
         return orderDateTime;
     }
 
-    public OrderSummary(Long orderId, int apples, int oranges, double totalPrice, LocalDateTime orderDateTime) {
+    public OrderSummary(Long orderId, int apples, int oranges, BigDecimal totalPrice, LocalDateTime orderDateTime) {
         this.orderId = orderId;
         this.apples = apples;
         this.oranges = oranges;
